pyproject.toml
[project]
name = "test-antigravity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "requests>=2.32.5",
]
